feat: add engine init module and rewrite db.ts as thin facade
